requirements.txt
Flask>=1.0
mysql-connector-python~=9.2.0
requests~=2.32.3
python-dotenv~=1.0.1
openai~=1.65.2